moved specific tests to data
