+++ packages/pymodaq_utils/tests/utils_test.py
@@ -1,5 +1,3 @@
-import random
-
 import numpy as np
 import os
 import pytest
@@ -9,7 +7,6 @@ import datetime
 
 from serializall import SerializableFactory
 
-import pymodaq_data.plotting.utils
 from pymodaq_utils import utils
 
 
@@ -287,38 +284,3 @@ class TestFindInIterable:
         assert selection[1] == objects[2], 2
 
 
-class TestPlotColor:
-
-    def test_iterable(self):
-        with pytest.raises(TypeError):
-            pcolor = pymodaq_data.plotting.utils.PlotColors((0, 0, 0))
-
-    def test_non_integer(self):
-        with pytest.raises(TypeError):
-            pcolor = pymodaq_data.plotting.utils.PlotColors([(0, 0., 0)])
-
-    def test_non_8bits(self):
-        with pytest.raises(TypeError):
-            pcolor = pymodaq_data.plotting.utils.PlotColors([(0, 256, 0)])
-
-    def test_negative(self):
-        with pytest.raises(TypeError):
-            pcolor = pymodaq_data.plotting.utils.PlotColors([(0, -5, 0)])
-
-    def test_get_item(self):
-        pcolor = pymodaq_data.plotting.utils.PlotColors()
-
-        for _ in range(10):
-            item = random.randrange(100)
-            color = pcolor[item]
-
-    def test_len(self):
-        N = random.randrange(10) + 1
-        pcolor = pymodaq_data.plotting.utils.PlotColors([(0, 0, 0) for _ in range(N)])
-        assert len(pcolor) == N
-
-    def test_iter(self):
-        pcolor = pymodaq_data.plotting.utils.PlotColors()
-
-        np.array(pcolor)
-        colors = [color for color in pcolor]
